Removed reference to 'Highlights"
As the sidebar highlight function is removed, this small update simply removes mentions of the feature from the page.  I also added wordage to advise on the matter of possible screen size issues.

"The feature is made to work with the user’s native screen size.  Note, that on some smaller screens the sidebar may place itself in a fixed position to accommodate the display area. "